Tunable CQL client creation (#4270)
* CQL factory interface, abstract helper and concrete impl

* Simple changes based on comments

* Simplified resource management.

* Added missing default
